Fix E2SM-RC Report Style 4 and add generic functions
This change fixes Report Style 4 issues regarding RRC connection status reported
to xApps.
Also adds some generic functions to build with RAN Parameters.
Fixes and organizes CMakeLists.txt files.

+++ e2sim/e2sm/rc/services/report_style4.hpp
@@ -89,7 +89,7 @@ public:
 
     bool generate_ueid_report_info(UEID_t &ueid, const std::string &imsi);
 
-    bool generate_ran_params_report_info(const e_RRC_State changed_to, const int rsrp, const int rsrq, const int sinr,
+    bool generate_ran_params_report_info(const e_RRC_State changed_to, const long rsrp, const long rsrq, const long sinr,
             std::vector<E2SM_RC_IndicationMessage_Format2_RANParameter_Item_t *> &params);
 
     void encode_and_send_report_msg(std::vector<common::rc::indication_msg_format2_ueid_t> &ue_ids);
